requirements.txt
# MedGemma Impact Challenge - Requirements
# Python 3.10+ required

# Core ML/AI
torch>=2.0.0
transformers>=4.50.0  # Required for MedGemma
accelerate>=0.27.0

# Data handling
Pillow>=10.0.0
numpy>=1.24.0

# Hugging Face
huggingface-hub>=0.20.0

# Demo app
gradio>=4.0.0

# Utilities
tqdm>=4.65.0

# Testing
pytest>=7.0.0

# Guidelines RAG
sentence-transformers>=2.2.0

# Edge AI (optional - for CPU-only inference)
onnx>=1.14.0
onnxruntime>=1.15.0
